Project gets more compleicated, now the drawing is done by a dedicated system. We are simulating a complex system, where RTTI has now forced us to have 3 different places. a) what a thing is b) what is done to thing c) logic that maps a thing is and what it does.
This is a good example of why some people argue vehamently against RTTI. We have cast away formal typing system, and now we are "leaking" information about the object outside of the class boundaries.
This sort of code is easy enough to write, but hard to maintain, undermining the point of Object-oriented coding.