Now the restructuring begins
Hold on I am going to try and walk you through a few steps.
First let's introduce the concept a policy. This allows us to seperate the concept of what a thing is and how it is treated.
At this stage, this concept simply adds more code, and changes nothing, but it allows us to start changing things