Now we are getting some where.
We can now clearly see that the ObjectType was never really needed. The RenderPolicy was a missing concept. Using it directly, we can see no longer need to know the type of the object, only it's policy.

Using a strongly-typed system has taught us to use the type to do things. However, this is only because the type is compiler checked. Once we move to run-time typing, it should no longer be the prefered method of doing things.

We can now add new objects without changing anything else. If we add new rendering policies we only have to change the enum and the renderer. This is how it should be. We only  touch both the renderer and the new class if we add a new class that needs a new policy.

But we can still improve this....